+++ Demo/demo01/ResourceEditWindow.cpp
@@ -56,8 +56,7 @@ void ResourceEditWindow::Load(const char* resource_file) {
   if (TBFile* file = TBFile::Open(m_resource_filename, TBFile::MODE_READ)) {
     TBTempBuffer buffer;
     if (buffer.Reserve(file->Size())) {
-      uint32_t size_read =
-          file->Read(buffer.GetData(), 1, buffer.GetCapacity());
+      size_t size_read = file->Read(buffer.GetData(), 1, buffer.GetCapacity());
       m_source_edit->SetText(buffer.GetData(), size_read);
     }
     delete file;
