+++ src/tb/tb_node_ref_tree.cpp
@@ -57,7 +57,7 @@ void TBNodeRefTree::InvokeChangeListenersInternal(const char* request) {
 }
 
 // static
-TBNodeRefTree* TBNodeRefTree::GetRefTree(const char* name, int name_len) {
+TBNodeRefTree* TBNodeRefTree::GetRefTree(const char* name, size_t name_len) {
   for (TBNodeRefTree* rt = s_ref_trees.GetFirst(); rt; rt = rt->GetNext())
     if (strncmp(rt->GetName(), name, name_len) == 0) return rt;
   return nullptr;
