+++ src/tb/tb_system.h
@@ -120,7 +120,7 @@ class TBFile {
   static TBFile* Open(const char* filename, TBFileMode mode);
 
   virtual ~TBFile() {}
-  virtual long Size() = 0;
+  virtual size_t Size() = 0;
   virtual size_t Read(void* buf, size_t elemSize, size_t count) = 0;
 };
